+++ src/torchjd/aggregation/__init__.py
@@ -3,7 +3,7 @@ from torchjd.aggregation.bases import Aggregator, WeightedAggregator, Weighting
 from torchjd.aggregation.cagrad import CAGradWeighting
 from torchjd.aggregation.constant import ConstantWeighting
 from torchjd.aggregation.dualproj import DualProjWrapper
-from torchjd.aggregation.graddrop import GradDropAggregator
+from torchjd.aggregation.graddrop import GradDrop
 from torchjd.aggregation.imtl_g import IMTLGWeighting
 from torchjd.aggregation.krum import KrumWeighting
 from torchjd.aggregation.mean import MeanWeighting